新增編輯用戶資料ＡＰＩ

+++ utils/cloudStorage-util.js
@@ -9,25 +9,33 @@ const deleteImages = async (imageUrls) => {
 
     // 檢查是否有圖片 URLs , -1 表示沒有圖片
     if (!imageUrls || imageUrls.length === 0) {
+      console.log("沒有圖片");
       return "-1";
     }
 
     // 使用 Promise.all 處理多個非同步請求
     await Promise.all(
       imageUrls.map(async (imageUrl) => {
-        // 透過正則表達式提取 "o=" 參數的內容，並進行 URL 解碼
-        const filePath = decodeURIComponent(imageUrl.match(/\/o\/([^?]+)/)[1]);
-        const file = bucket.file(filePath);
+        // 使用更可靠的方式提取檔案路徑
+        const match = imageUrl.match(/\/o\/(.*?)\?/);
 
-        const [exists] = await file.exists();
+        if (match) {
+          const filePath = decodeURIComponent(match[1]); // 將提取到的路徑進行 URL 解碼
+          const file = bucket.file(filePath);
 
-        if (exists) {
-          await file.delete(); // 刪除檔案
+          const [exists] = await file.exists();
+
+          if (exists) {
+            await file.delete(); // 刪除檔案
+          }
+          console.log("成功刪除檔案");
+        } else {
+          console.log(`無法從 URL 提取檔案路徑: ${imageUrl}`);
         }
       })
     );
   } catch (e) {
-    console.error("刪除圖片作業有問題:", e);
+    console.log("刪除圖片作業有問題:", e);
   }
 };
